+++ docs/release notes.rst
@@ -5,5 +5,9 @@ Release Notes
 v1.0
 ----
 
+**Features:**
+  * Close #28: The string representation of a category now includes
+    information about its parent category.
+
 **Documentation:**
   * Close #8: Create initial documentation.